feat(druck): support various content processors
A content processor is a function(content, page) that
processes the content and returns the processed result.

Multiple processors can now be defined in the configuration
via the contentProcessors property. It may be an object or fn.

```
{
  contentProcessors: {
    '*.md': function(content, page) { return marked(content); },
    '*': function(content, page) { return emoji(content); },
  }
}
```

Page ids are matched against the pattern; processor order is
preserved, so that multiple processors are used in order.

```
{
  contentProcessors: function(page) {
    // return applicable processors, single processor or none
    return [] || fn;
  }
}
```

Content processing can also be disabled entirely by configuring
contentProcessors=false.

Content will simply be kept as is, if no processors are defined.

Closes #4

BREAKING CHANGE:

* the method `Kartoffeldruck.isMarkdown` is not needed anymore and
  removed from the +public+ API.